% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_label_analysis_utils.r
\name{.cdh_save_plot}
\alias{.cdh_save_plot}
\title{Save a ggplot to a directory if ggplot2/the plot are available}
\usage{
.cdh_save_plot(plot, dir_path, filename, width = 7, height = 5)
}
\description{
Save a ggplot to a directory if ggplot2/the plot are available
}
