% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_d2t_analysis.r
\name{ds.add_cohort_column}
\alias{ds.add_cohort_column}
\title{Derive a cohort column from pat_ID prefix on a server-side object}
\usage{
ds.add_cohort_column(
  df,
  pat_id_col = "pat_ID",
  prefix_regex = "^[A-Za-z]+",
  newobj = NULL,
  datasources = NULL
)
}
\description{
Derive a cohort column from pat_ID prefix on a server-side object
}
