% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_build_survival_table.r
\name{ds.build_survival_table}
\alias{ds.build_survival_table}
\title{Build a patient-level survival table on the server}
\usage{
ds.build_survival_table(
  df,
  pat_id_col = "pat_ID",
  event_col = "D2T",
  time_col = "Visit_months_from_diagnosis",
  entry_time_col = NULL,
  group_col = NULL,
  covariate_cols = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\description{
Build a patient-level survival table on the server
}
