% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_btsdmard.r
\name{ds.check_btsdmard_presence}
\alias{ds.check_btsdmard_presence}
\title{Check whether at least one of bDMARD/tsDMARD is present, per row}
\usage{
ds.check_btsdmard_presence(
  df,
  bDMARD_col = "bDMARD",
  tsDMARD_col = "tsDMARD",
  nfilter = 5,
  datasources = NULL
)
}
\description{
Check whether at least one of bDMARD/tsDMARD is present, per row
}
