% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_check_types.r
\name{ds.check_types}
\alias{ds.check_types}
\title{Check variable types on a server-side data frame}
\usage{
ds.check_types(df, datasources = NULL)
}
\arguments{
\item{df}{A character string specifying the name of the server-side data frame
containing numeric and categorical variables to be checked.}

\item{datasources}{A list of \code{\link[DSI]{DSConnection-class}} objects obtained
after login. If the \code{datasources} argument is not specified, the default set of
connections will be used: see \code{\link[DSI]{datashield.connections_default}}.}
}
\value{
Invisible \code{NULL}. Messages are printed to indicate which columns,
  if any, are non-compliant on each server.
}
\description{
Verify that numeric and categorical variables in a server-side data frame
  comply with predefined valid ranges and values. Uses the server-side functions
  \code{check_numericDS} and \code{check_categoricalDS} to identify non-compliant columns.
  Messages are printed for each server indicating which columns, if any, are invalid.

Server functions called: \code{check_numericDS}, \code{check_categoricalDS}
}
