% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_compute_derived_variables.r
\name{ds.compute_derived_variables}
\alias{ds.compute_derived_variables}
\title{Compute derived clinical variables (USPDGS, DAS2C, CDAI50, SDAI)}
\usage{
ds.compute_derived_variables(df, newobj = NULL, datasources = NULL)
}
\description{
Compute derived clinical variables (USPDGS, DAS2C, CDAI50, SDAI)
}
