% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_d2t_analysis.r
\name{ds.d2t_between_cohort_test}
\alias{ds.d2t_between_cohort_test}
\title{Objective 2: between-cohort heterogeneity (Kruskal-Wallis + chi-square)}
\usage{
ds.d2t_between_cohort_test(
  df,
  numeric_vars = NULL,
  categorical_vars = NULL,
  group_col = "cohort",
  nfilter = 5,
  datasources = NULL
)
}
\description{
Objective 2: between-cohort heterogeneity (Kruskal-Wallis + chi-square)
}
