% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_d2t_analysis.r
\name{ds.d2t_followup_stats}
\alias{ds.d2t_followup_stats}
\title{Objective 3: follow-up duration heterogeneity by cohort}
\usage{
ds.d2t_followup_stats(
  df,
  pat_id_col = "pat_ID",
  visit_months_col = "Visit_months_from_diagnosis",
  group_col = "cohort",
  min_months = 24,
  nfilter = 5,
  datasources = NULL
)
}
\description{
Objective 3: follow-up duration heterogeneity by cohort
}
