% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_d2t_analysis.r
\name{ds.d2t_incidence}
\alias{ds.d2t_incidence}
\title{Objective 5: D2T incidence -- overall, by cohort, by diagnosis-year
  bin, by disease-duration bin (NOT time-to-event -- see Objective 6 note)}
\usage{
ds.d2t_incidence(
  df,
  event_col = "D2T",
  pat_id_col = "pat_ID",
  visit_months_col = "Visit_months_from_diagnosis",
  year_diagnosis_col = "Year_diagnosis",
  group_col = "cohort",
  year_bin_width = 2,
  duration_bin_width = 12,
  nfilter = 5,
  datasources = NULL
)
}
\description{
Objective 5: D2T incidence -- overall, by cohort, by diagnosis-year
  bin, by disease-duration bin (NOT time-to-event -- see Objective 6 note)
}
