% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_d2t_analysis.r
\name{ds.d2t_visit_interval_stats}
\alias{ds.d2t_visit_interval_stats}
\title{Objective 4: inter-visit interval heterogeneity by cohort}
\usage{
ds.d2t_visit_interval_stats(
  df,
  pat_id_col = "pat_ID",
  visit_months_col = "Visit_months_from_diagnosis",
  group_col = "cohort",
  nfilter = 5,
  datasources = NULL
)
}
\description{
Objective 4: inter-visit interval heterogeneity by cohort
}
