% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_fill_missing_data.r
\name{ds.fill_missing_data}
\alias{ds.fill_missing_data}
\title{Fill missing visit data on a server-side data frame}
\usage{
ds.fill_missing_data(
  df,
  pat_id_col,
  visit_col,
  value_col,
  filled_newobj = NULL,
  datasources = NULL
)
}
\description{
Fill missing visit data on a server-side data frame
}
