% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_generic_dataframe_ops.r
\name{ds.select_columns}
\alias{ds.select_columns}
\title{Select (keep) specific columns on a server-side data frame}
\usage{
ds.select_columns(df, columns, newobj = NULL, datasources = NULL)
}
\description{
Select (keep) specific columns on a server-side data frame
}
