% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_generic_dataframe_ops.r
\name{ds.set_column_value}
\alias{ds.set_column_value}
\title{Set a column's value on a server-side data frame}
\usage{
ds.set_column_value(
  df,
  column,
  value = NULL,
  value_expr = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\description{
Set a column's value on a server-side data frame
}
