% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_cast_to_factor.r
\name{ds.cast_to_factor}
\alias{ds.cast_to_factor}
\title{Cast specified columns to factor on a server-side data frame}
\usage{
ds.cast_to_factor(df, columns, modified_obj = NULL, datasources = NULL)
}
\description{
Cast specified columns to factor on a server-side data frame
}
