% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dummy_encode_ordinal.r
\name{ds.dummy_encode_ordinal}
\alias{ds.dummy_encode_ordinal}
\title{Dummy-encode ordinal DMARD-style variables on a server-side data frame}
\usage{
ds.dummy_encode_ordinal(
  df,
  columns = c("csDMARD1", "csDMARD2", "csDMARD3", "bDMARD", "tsDMARD", "GC_type", "GC"),
  newobj = NULL,
  datasources = NULL
)
}
\description{
Dummy-encode ordinal DMARD-style variables on a server-side data frame
}
