% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_generic_dataframe_ops.r
\name{ds.filter_rows}
\alias{ds.filter_rows}
\title{Filter rows of a server-side data frame by a simple condition}
\usage{
ds.filter_rows(
  df,
  column,
  operator = c("==", "!=", ">", "<", ">=", "<=", "\%in\%", "is.na", "not.na"),
  value = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\description{
Filter rows of a server-side data frame by a simple condition
}
