% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_force_convert_types.r
\name{ds.force_convert_types}
\alias{ds.force_convert_types}
\title{Force type conversion on a server-side data frame}
\usage{
ds.force_convert_types(
  df,
  numeric_cols = NULL,
  categorical_cols = NULL,
  newobj = NULL,
  check_zero_anomalies = TRUE,
  datasources = NULL
)
}
\description{
Force type conversion on a server-side data frame
}
