% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_global_imputation.R
\name{ds.global_imputation}
\alias{ds.global_imputation}
\title{Global imputation of a server-side object using MICE}
\usage{
ds.global_imputation(
  df,
  id_col = NULL,
  m = 5,
  maxit = 30,
  imputed_newobj = NULL,
  datasources = NULL
)
}
\description{
Global imputation of a server-side object using MICE
}
