% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_impute_missing_data.r
\name{ds.impute_missing_data}
\alias{ds.impute_missing_data}
\title{Longitudinal (patient-grouped) MICE imputation -- preferred method}
\usage{
ds.impute_missing_data(
  df,
  pat_id_col = "pat_ID",
  visit_col = "Visit",
  exclude_cols = NULL,
  m = 5,
  maxit = 30,
  nfilter = 5,
  imputed_newobj = NULL,
  datasources = NULL
)
}
\description{
Longitudinal (patient-grouped) MICE imputation -- preferred method
}
