% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_remove_columns.r
\name{ds.remove_columns}
\alias{ds.remove_columns}
\title{Remove specified columns from a server-side data frame}
\usage{
ds.remove_columns(df, col_names, newobj = NULL, datasources = NULL)
}
\description{
Remove specified columns from a server-side data frame
}
